+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### Added
 
+- `AdvertisingConfig.shouldLoadAdItem` callback to filter ad items before they are loaded
 - iOS: `PictureInPictureConfig.shouldExitOnForeground` to automatically exit PiP mode when the app transitions to the foreground
 
 ### Changed
 
 - Update Bitmovin's native Android SDK version to `3.150.0+jason`
 - Update Bitmovin's native iOS SDK version to `3.111.1`
+- Android: Kotlin version to `2.2.20`
 
 ## [1.17.0] - 2026-04-17